+++ client/gulpfile.js
@@ -1,19 +1,15 @@
-// load modules
-var gulp = require('gulp');
-var del = require('del');
-var path = require('path');
-var runSequence = require('run-sequence');
-var plumber = require('gulp-plumber');
-var concat = require('gulp-concat');
-var stripDebug = require('gulp-strip-debug');
-var uglify = require('gulp-uglify');
-var sourcemaps = require('gulp-sourcemaps');
+const { src, dest, watch, series } = require('gulp');
+const del = require('del');
+const path = require('path');
+const plumber = require('gulp-plumber');
+const sourcemaps = require('gulp-sourcemaps');
+const concat = require('gulp-concat');
+const stripDebug = require('gulp-strip-debug');
+const uglify = require('gulp-uglify');
 
-// load paths
-var paths = {
+const paths = {
 	"src": "./src/",
 	"dist": "./dist/",
-	"webroot": "../../",
 	
 	"scripts": {
 		"src": "javascript",
@@ -22,44 +18,66 @@ var paths = {
 	}
 };
 
-gulp.task('scripts', ['cleanscripts'], function() {
-	return gulp
-		.src(paths.src + paths.scripts.src + paths.scripts.filter)
+const script_builds = {
+	"jquery.min.js": [
+		"node_modules/jquery/dist/jquery.js",
+	],
+	
+	"jquery-validate.min.js": [
+		"node_modules/jquery-validation/dist/jquery.validate.js",
+	],
+	
+	"mailchimp-validation.min.js": [
+		"src/javascript/mailchimp-validation.js"
+	]
+};
+
+function buildScripts(cb) {
+	var scriptNames = Object.keys(script_builds);
+	scriptNames.forEach(function(scriptName) {
+		src(
+			script_builds[scriptName],
+			{
+				cwd: path.join(process.cwd(), './'),
+				nosort: true
+			}
+		)
 		.pipe(plumber({
 			errorHandler: onError
 		}))
 		.pipe(sourcemaps.init())
-		.pipe(concat('mailchimp-validation.min.js'))
+		.pipe(concat(scriptName))
 		.pipe(stripDebug())
 		.pipe(uglify({mangle: false}))
-	    .pipe(sourcemaps.write('./'))
-		.pipe(gulp.dest(paths.dist + paths.scripts.dist));
-});
+		.pipe(sourcemaps.write('./'))
+		.pipe(dest(paths.dist + paths.scripts.dist));
+	});
+	cb();
+}
 
-gulp.task('cleanscripts', function() {
-	return del.sync([
-		paths.dist + paths.scripts.dist
+function cleanScripts(cb) {
+	del([
+		paths.dist + paths.scripts.dist + "*.(js|map)"
 	]);
-});
+	cb();
+}
 
-gulp.task('watch', function() {
-	gulp.watch(paths.src + paths.scripts.src + paths.scripts.filter, ['scripts']);
-});
+function watchScripts() {
+	// watch for script changes
+	watch(paths.src + paths.scripts.src + paths.scripts.filter, series(cleanScripts, buildScripts));
+}
 
-gulp.task('build', function (callback) {
-	runSequence(
-		['scripts'],
-	    callback
-	)
-});
-
-gulp.task('default', function (callback) {
-	runSequence(
-		['scripts', 'watch'],
-		callback
-	)
-});
-
-var onError = function(err) {
+function onError(err) {
     console.log(err);
 }
+
+exports.build = series(
+	cleanScripts,
+	buildScripts,
+);
+
+exports.default = series(
+	cleanScripts,
+	buildScripts,
+	watchScripts
+);